+++ youtugo-domain/src/main/java/org/nexters/az/post/exception/NonExistentBookMarkException.java
@@ -1,0 +1,10 @@
+package org.nexters.az.post.exception;
+
+import org.nexters.az.exception.NotFoundException;
+
+public class NonExistentBookMarkException extends NotFoundException {
+    public NonExistentBookMarkException() {
+        this("Post does not exist");
+    }
+    public NonExistentBookMarkException(String message) {super(message);}
+}
